+++ Sources/QTFeatures.m
@@ -174,9 +174,6 @@ void QTInstallFeatures(void) {
 // BEGIN 0.13 AD PROFILE
     QTInstallAdProfile();
 // END 0.13 AD PROFILE
-// BEGIN 0.10 PLAYER PROBE
-    QTInstallPlayerProbe();
-// END 0.10 PLAYER PROBE
     if (QTOn(@"feedAds") || QTOn(@"shorts") || (QTOn(@"extendedFeed") && (QTOn(@"playables") || QTOn(@"eventPromos") || QTOn(@"topicsShelves") || QTOn(@"edgeCards") || QTOn(@"inspectElements") || QTOn(@"mixes") || QTOn(@"watchAgain")))) {
         QTHook(@"YTInnerTubeCollectionViewController",@"addSectionsFromArray:",@"v@",^id(IMP old,SEL sel) {
             return ^(id object,NSArray *sections) {